moves network Magic type from zebra-network to zebra-chain
