+++ zebra-chain/src/parameters.rs
@@ -22,7 +22,7 @@ mod transaction;
 pub mod arbitrary;
 
 pub use genesis::*;
-pub use network::{testnet, Network, NetworkKind};
+pub use network::{magic::Magic, testnet, Network, NetworkKind};
 pub use network_upgrade::*;
 pub use transaction::*;
